added likes for all posts
